+++ src/main/java/domain/table/Tables.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.table;
 
 import java.util.Collections;
 import java.util.List;
