+++ src/main/java/view/OutputView.java
@@ -3,8 +3,8 @@ package view;
 import java.util.List;
 import java.util.Map;
 
-import domain.Menu;
-import domain.Table;
+import domain.menu.Menu;
+import domain.table.Table;
 
 public class OutputView {
 	private static final String TOP_LINE = "┌ ─ ┐";
