+++ src/test/java/domain/MenuFixture.java
@@ -1,5 +1,8 @@
 package domain;
 
+import domain.menu.Menu;
+import domain.menu.MenuRepository;
+
 public class MenuFixture {
 	static final Menu FRIED = MenuRepository.menus().get(0);
 	static final Menu SEASONED = MenuRepository.menus().get(1);
