+++ src/test/java/domain/TableTest.java
@@ -3,6 +3,7 @@ package domain;
 import static domain.MenuFixture.*;
 import static org.assertj.core.api.Assertions.*;
 
+import domain.table.Table;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
